Move MessageInput view, style, controller to standalone package
