+++ sample/src/main/res/layout/fragment_channel.xml
@@ -34,7 +34,7 @@
         app:layout_constraintStart_toStartOf="parent"
         app:layout_constraintTop_toTopOf="parent" />
 
-    <com.getstream.sdk.chat.view.MessageInputView
+    <com.getstream.sdk.chat.view.messageinput.MessageInputView
         android:id="@+id/messageInputView"
         android:layout_width="match_parent"
         android:layout_height="wrap_content"
